PLAT-69761: Merge all ConnectSDK code changes related to SWMD in Github
